feat: create tokenomics module.
logion-network/logion-internal#1283
